chore: rename and re-org files, update tailwind colors and buttons for banner (complete) and carousel (wip)
