+++ components/badge.tsx
@@ -1,7 +1,7 @@
 import { type VariantProps, cva } from "class-variance-authority";
 import * as React from "react";
 
-import { cn } from "../utils";
+import { cn } from "./utils";
 
 const badgeVariants = cva(
   "inline-flex items-center rounded-md border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2",
@@ -9,12 +9,11 @@ const badgeVariants = cva(
     variants: {
       variant: {
         default:
-          "border-transparent bg-primary text-primary-foreground shadow hover:bg-primary/80",
+          "inline-flex items-center rounded-full bg-yellow-500 px-2 py-1 text-xs font-medium text-gray-600 ring-1 ring-gray-500/10 ring-inset",
         secondary:
-          "border-transparent bg-secondary text-secondary-foreground hover:bg-secondary/80",
-        destructive:
-          "border-transparent bg-destructive text-destructive-foreground shadow hover:bg-destructive/80",
-        outline: "text-foreground",
+          "inline-flex items-center rounded-full bg-secondary-blue-500 px-2 py-1 text-xs font-medium text-white ring-1 ring-gray-500/10 ring-inset",
+        outline: 
+        "inline-flex items-center rounded-full border border-input bg-secondary-blue-500 px-2 py-1 text-xs font-medium text-white ring-1 ring-gray-500/10 ring-inset",
       },
     },
     defaultVariants: {
@@ -29,7 +28,7 @@ export interface BadgeProps
 
 function Badge({ className, variant, ...props }: BadgeProps) {
   return (
-    <div className={cn(badgeVariants({ variant }), className)} {...props} />
+    <span className={cn(badgeVariants({ variant }), className)} {...props} />
   );
 }
