+++ components/header/ccv-status-utils.js
